Extracted all constants to const
