+++ js/const/comment-avatar-number.js
@@ -1,0 +1,6 @@
+const CommentAvatarNumber = {
+  MIN: 1,
+  MAX: 6,
+};
+
+export default CommentAvatarNumber;
