+++ js/const/commentator-names.js
@@ -1,0 +1,12 @@
+const COMMENTATOR_NAMES = [
+  'Иван',
+  'Хуан Себастьян',
+  'Мария',
+  'Кристоф',
+  'Виктор',
+  'Юлия',
+  'Люпита',
+  'Вашингтон',
+];
+
+export default COMMENTATOR_NAMES;
