+++ js/const/index.js
@@ -1,0 +1,7 @@
+export {default as PHOTOS_COUNT} from './photos-count.js';
+export {default as MAX_COMMENT_NUMBER} from './max-comment-number.js';
+export {default as COMMENT_MESSAGES} from './comment-messages.js';
+export {default as COMMENTATOR_NAMES} from './commentator-names.js';
+export {default as CommentAvatarNumber} from './comment-avatar-number.js';
+export {default as PhotoLikesCount} from './photo-likes-count.js';
+export {default as PHOTO_DESCRIPTIONS} from './photo-descriptions.js';
