+++ js/const/max-comment-number.js
@@ -1,0 +1,3 @@
+const MAX_COMMENT_NUMBER = 10;
+
+export default MAX_COMMENT_NUMBER;
