+++ js/const/photo-descriptions.js
@@ -1,0 +1,12 @@
+const PHOTO_DESCRIPTIONS = [
+  'Description 1',
+  'Description 2',
+  'Description 3',
+  'Description 4',
+  'Description 5',
+  'Description 6',
+  'Description 7',
+  'Description 8',
+];
+
+export default PHOTO_DESCRIPTIONS;
