+++ js/const/photo-likes-count.js
@@ -1,0 +1,6 @@
+const PhotoLikesCount = {
+  MIN: 15,
+  MAX: 200,
+};
+
+export default PhotoLikesCount;
