+++ js/const/photos-count.js
@@ -1,0 +1,3 @@
+const PHOTOS_COUNT = 25;
+
+export default PHOTOS_COUNT;
