+++ js/data/create-comment.js
@@ -1,29 +1,5 @@
 import {uniqueId, getRandomArrayElement, getRandomInteger} from '../utils/index.js';
-
-const COMMENT_MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
-];
-
-const COMMENTATOR_NAMES = [
-  'Иван',
-  'Хуан Себастьян',
-  'Мария',
-  'Кристоф',
-  'Виктор',
-  'Юлия',
-  'Люпита',
-  'Вашингтон',
-];
-
-const CommentAvatarNumber = {
-  MIN: 1,
-  MAX: 6,
-};
+import {COMMENT_MESSAGES, COMMENTATOR_NAMES, CommentAvatarNumber} from '../const/index.js';
 
 const commentIdGenerator = uniqueId();
